chore: chaning layout
